✨(back) add transcode_pipeline property for VOD
This property is used to store the pipeline used to transcode
a VOD video. Currently, every VOD video is transcoded using the
same pipeline (AWS), but this will change in the future.
This property is then used to build the video's URL that will
be used to play the video in the client.