🛂(frontend) remove persistance student token
Opening a classroom invite link switch our current token to
a student token, so when we go back to the browser instructor tab,
the resources cannot be access anymore, this is because of the
persistency of the student token. We remove this persistency,
the student token is now only valid for the current tab.