♻️(back) refacto delete videos from S3
The way we handle the "delete" videos on AWS is not optimal and does
not work on the newly added storage Scaleway. To simplify the process,
we will move the deleted files into a deleted directory that will have
one lifecycle rule that will delete the files after 21 days. The "move"
can takes some times so we want it to be run by the celery task queue.