♻️(lib_video) manage sharing error when resource is already shared
When a user tried to share a resource when another
one was already shared, the sharing was not started and an
error was displayed.
We now stop the sharing of the first resource and start the sharing
of the new one if we get the error 'Video is already sharing' from
the backend.