✨(back) add scaleway object storage support
Add support for scaleway object storage. It is a
drop-in replacement for S3, that is boto3 compatible.
Use django-storage to create a class that will be used to
upload videos. Refactor the s3 utils to either use the
aws client or the scaleway client.
Also add terraform code to create the bucket and the update
the cloudfront settings.