⚡️(lti) separate css from the js bundle
We need css files and font from the Cunningham packages
to style our app, but the css was included directly in the js bundle,
which was not optimal.
The css is now in a separate file, which is loaded in the html
template. This allows the browser to cache the css file and
improves the loading time of the page.