🚨(standalone) fix build warning duplicate module
In the generics.ts module, importing the module with await import was
creating a chunk of the module imported, but Vite was pinpointing
that the module was already bundled in the main chunk via index.
This commit fixes the warning by importing the module with a
standard import, it doesn't have to be a promise anymore.