🐛(standalone) fix vtt.js error
vtt.js expect VTTRegion to be already defined, but it is not the case,
it seems that the Window object didn't have the time to load
VTTRegion before vtt.js try to use it. By importing vtt.js in the
component, we make sure that Window had the time to load VTTRegion.