✅(standalone) add jest and babel config files
All our tests are already written in jest, so we
will use Jest with Vite.
CRA was managing the jest config for us, but now we have to do
it ourselves, to do so we need to add a jest config file, and the
babel config file. We added some babel plugins to the babel config
to make it work with jest and vite.