✅(standalone) manage svg like CRA
CRA with Jest has a special treatment for SVG files. It uses a Jest
transformer to convert SVG files into a svg with some basic
information. It is quite usefull to test wich svg is loaded in
our tests. In order to have the same behavior in our Vite standalone,
we need to add the same transformer.
We had to fix another problem, the svg in our app are loaded with
a query string. The transformer was not able to handle it. We had to
add a babel plugin (babel-plugin-import-remove-resource-query)
to remove the query string before the transformer
is called.