🔧(tray) configure tray to handle connection with peertube runner
In order to manage connection with the peertube runner, we must have to
configure the tray. First a new location is added in nginx to manage the
socket connection. This connection is made between service directly
without going out of the k8s cluster, so the service name must be added
to the django allowed hosts.