📌(back) pin lxml to version < 5
We are not able tu use lxml >= 5.0.0. We have a segfault when using it.
The solution could be to install it without binaries but it is not
possible to do it using the setup.cfg file.
We have to investigate if it will be possible if we switch to pyproject.

Linked issues :

https://github.com/SAML-Toolkits/python3-saml/issues/388
https://github.com/SAML-Toolkits/python3-saml/issues/389