🐛(back) correctly send xapi statement to c configured LRS
Since xAPI statements are build using a dedicated class, they return a
dict directly usable as a statement. The XAPI.send method was still
using the `get_statement` method that does not exists anymore. To fix
this issue we can directly use the statement property set in the send
method.