🐛(back) correctly set startime with filtering medialive logs
Medialive has changed when the alerts are set for a medialive channels.
Before alerts were set with a delay once the channel started, now they
are set at the exact same time. So we have to adjust the startTime when
filtering logs in the check_live_state command to be sure to have all
the logs from the beginning. Otherwise some logs are missed and the
command is failing.