✨(front) persist username on classroom join
Currently, each time an anonymous user wants to join a classroom, it has to
provide a user name. It orders to improve UX, we decide to store username
provided in local storage when user join the classroom. In this way, next time
it wants to join, the input will be prefilled.