🐛(back) manage playlist constraint with DRF 3.15
In DRF 3.15 they added, in the serializer validator, the model
constraint validation. With the playlist model, we have to force the
lti_id default value and manage better in the test the constraint
between organization or consumer_site existence with the lti_id.

See https://github.com/encode/django-rest-framework/pull/7438