♻️(back) compute hls url based on video resolutions available
the combine usage of peertube-runner and cloudfront can lead to
situation where the master HLS url is cached for 24 hours without all
the resolutions available. Once the first resolution is transcoded, the
master HLS file is generated and available in Marsha. The websocket
updates the video object and the file is in cache for 24 hours. We
introduce a query parameter used to compute the resolutions sha256. So
every time the list of resolutions available is changing, this hash also
changes and a new cache is generated.