📌(back) pin django-storages to version 1.14.3
In version 1.14.4 there are at least two modifications made that lead to
a breaking change in Marsha.
The most annoying one is linked to this issue:
jschneier/django-storages#1430 and we have to
wait a newer version with a fix to have the previous behaviour. This fix
is related to a security issue in django. This security is fixed in
version 4.2.14 and we already use this version, so we are safe.
The second one is related to how the signature in computed when an url
is generated. Previously the signature was generated no matter if we
need it or not and then we choose to remove the signautre part using the
private method `_strip_signing_parameters`. This private does not exists
anymore, instead a new setting is used, we have to set the setting
`querystring_auth` to False to not compute the signature, it's real
improvement as it saves the cost of computing the signature.