♻️(tray) split websocket and asgi apps in tray
We have detected a memory leak in the websocket application. An issue
exists on the django channels github repo about this topic.
We also detected that this memory leak can have side effect on the asgi
application, the application can become slower leading to timeout. To
remove this side effect we decided to create a deploy dedicated to the
websocket application and nginx is reponsible to use the right backend
based on the request path.

https://github.com/django/channels/issues/2094