⚗️(front) remove websocket connection in VOD student dashboard
in the VOD student dashboard we are using the websocket to update the
video view. Opening a websocket can consume lot of resources on the
backend application for very few user experience improvement. We want to
try to remove this connection to see if there is an impact on the
backend load vs degrading a little bit the user experience.