✨(front) add new store useFlags storing info about feature flags
We want to use feature flags in the lib-video. For that we have to
refactor how the flags are stored. For now they are in the useAppConfig
store that is only used in the lti_site app and can't be used in the
standalone site. A first step is to create a dedicated store storing the
flags and determine if a flag is enabled or not.