🐛(xAPI) use domain from lti jwt token when no consumer site used
When a xAPI request is made from a LTI context but the video was created
on the website, there is no consumer_site attached to the playlist. We
are using the domain from the consumer_site when the xAPI request is
made in a LTI context. To fix this issue, in the xAPI endpoint is
testing if there is a consumer site and if not the one from the LTI
passport is used.