🐛(front) enable join classroom button with local storage username
Currently, in the form to join a classroom, if the input is populated with the
username stored into local storage, the join button remains disabled until the
user changes the input value.
The user should be able to click immediately on the join button once a value
populates the field.