⬆️(back) upgrade to pytest 8.3.3
We are not using pytest 8 because it was not compatible with
pytest-playwright. It is not the case anymore, we can upgrade both. Once
upgraded there was a regression in a test about social auth backends
used in a middleware. The backend is replaced using the
override_settings decorator but the backend is put in cache earlier. We
have to reload the module social_core.backends.utils responible to build
this cache.