🔧(tray) increase connection timeout for peertube runner requests
Some requests to the `success` endpoint ends up returning a 504 Timeout, usually
for the bigger vod files.
The `success` endpoint loads the converted video file and manipulates it, which
could be taking some time.
Increasing the connection timeout on the Nginx conf to 300s so it does not
timeout due to longer files.