✨(scw) add SCW provisioning with Terraform
This Terraform project aims to replace the AWS terraform projet.
The state is managed by each user, and is stored on Scaleway S3.
The projet goal is to create all the Scaleway resources necessary
for Marsha to upload and distribute videos using several Scaleway services:
S3, serverless functions, edge services (when provider will be available)...